1:36

+++ maxSalary.java
@@ -49,9 +49,10 @@ public class MaxSalary {
         }
     } // The Reducer
 
-    public static class Reduce extends MapReduceBase implements Reducer<Text, IntWritable, Text, IntWritable> {
-        // public void reduce(Text key, Iterator<IntWritable> values, OutputCollector<Text, IntWritable> collector,
-        public void reduce(Text key, Iterable<DoubleWritable> values, OutputCollector<Text, IntWritable> collector,
+    public static class Reduce extends MapReduceBase implements Reducer<Text, DoubleWritable, Text, DoubleWritable> {
+        // public void reduce(Text key, Iterator<IntWritable> values,
+        // OutputCollector<Text, IntWritable> collector,
+        public void reduce(Text key, Iterable<DoubleWritable> values, OutputCollector<Text, DoubleWritable> collector,
                 Reporter reporter) throws IOException {
             // Text key2 = key;
             // int count = 0;
